+++ src/amCacheActions.js
@@ -1,5 +1,6 @@
 import { Http } from '@app-masters/js-lib';
 import AMCache from './cache';
+import get from 'lodash/get';
 
 class AMCacheActions {
     constructor (config) {
@@ -194,6 +195,7 @@ class AMCacheActions {
 
     // cache ok
     getObjects = (sort, populate, filter) => {
+
         return (dispatch) => {
             try {
                 this.setLoading(dispatch, true, this.config.cacheStrategy !== 'Online', this.config.cacheStrategy !== 'Cache');
@@ -213,7 +215,8 @@ class AMCacheActions {
                     // console.log("getObjects.sempreRetornar");
 
                     // When using cache, always SHOW cache data, even when getting from API first
-                    let replaceAll = fromCache === false && !filter && this.config.cacheStrategy !== 'CacheOnline';
+                    const persistCache = !this.config.persistCache && this.config.cacheStrategy === 'CacheOnline'; // Cache is the truth
+                    let replaceAll = fromCache === false && !filter && persistCache;
 
                     this.dispatchGetObjects(dispatch, response, replaceAll, filter);
                 };
@@ -268,7 +271,13 @@ class AMCacheActions {
         }
         if (filter)
             url += '&' + filter;
-        return Http.get(url);
+        return Http.get(url).then(response => {
+            if(this.config.nestedKey){
+                return get(response, this.config.nestedKey);
+            } else {
+                return response;
+            }
+        });
     }
 
     /* GET OBJECTS */
@@ -319,7 +328,13 @@ class AMCacheActions {
             populate = (populate ? populate : this.config.defaultPopulate);
             url += '?populate=' + populate;
         }
-        return Http.get(url);
+        return Http.get(url).then(response => {
+            if(this.config.nestedKey){
+                return get(response, this.config.nestedKey);
+            } else {
+                return response;
+            }
+        });
     }
 
     dispatchGetObject (dispatch, response, fromCache) {
@@ -423,7 +438,14 @@ class AMCacheActions {
     };
 
     _createObject = (input) => {
-        return Http.post(this.config.endPoint, input);
+        const sufix = this.config.createSufix || '';
+        return Http.post(this.config.endPoint + sufix, input).then(response => {
+            if(this.config.nestedKey){
+                return get(response, this.config.nestedKey);
+            } else {
+                return response;
+            }
+        });
     };
 
     /* UPDATE */
@@ -476,7 +498,14 @@ class AMCacheActions {
     };
 
     _updateObject (id, input) {
-        return Http.put(this.config.endPoint + id, input);
+        const sufix = this.config.updateSufix || '';
+        return Http.put(this.config.endPoint + id + sufix, input).then(response => {
+            if(this.config.nestedKey){
+                return get(response, this.config.nestedKey);
+            } else {
+                return response;
+            }
+        });
     }
 
     dispatchSaveObject (dispatch, response, secondaryAction) {
@@ -578,7 +607,14 @@ class AMCacheActions {
     };
 
     _deleteObject (id) {
-        return Http.delete(this.config.endPoint + id);
+        const sufix = this.config.deleteSufix || '';
+        return Http.delete(this.config.endPoint + id + sufix).then(response => {
+            if(this.config.nestedKey){
+                return get(response, this.config.nestedKey);
+            } else {
+                return response;
+            }
+        });
     }
 
     _dispatchDeleteObject (dispatch, id) {
@@ -592,6 +628,12 @@ class AMCacheActions {
         }
 
     }
+
+    resetObjects = () => {
+        return (dispatch) => {
+            dispatch({type: this.type('GET_OBJECTS'), payload: []})
+        };
+    };
 
     /* OTHER ACTIONS */