+++ src/cache.js
@@ -43,6 +43,14 @@ class AMCache {
 
     static merge (cache, objects) {
         // Validate - objects are array? objects have _id on root element?
+        if (objects && objects.length > 0) {
+            objects = objects.map(object => {
+                if (Object.keys(object).length > 0 && !object._id) {
+                    return {_id: 'fake_' + new Date().getTime(), ...object};
+                }
+                return object;
+            });
+        }
 
         // console.log('cache', cache);
         if (!cache || cache.length === 0) {
@@ -53,11 +61,6 @@ class AMCache {
 
             // Walk throught objects merging by id
             objects.map(object => {
-                if (!object._id) {
-                    object._id = 'fake_' + new Date().getTime();
-                    // console.log('Id created', object._id);
-                }
-
                 // find on cache
                 let index = cache.findIndex(item => {
                     return item._id === object._id;